+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [unreleased]
+### Fixes
+- [198](https://github.com/tobiasschuerg/InfluxDB-Client-for-Arduino/pull/198) - Effective passing Point by value
+
 ##  3.12.1 [2022-08-29]
 ### Fixes
 - [193](https://github.com/tobiasschuerg/InfluxDB-Client-for-Arduino/pull/193) - Automatically adjusting point timestamp  according to the setting of write precision.